LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.5_Tokens/7.5.2_Keywords/S7.5.2_A1.5-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected keyword 'default'. Expected a closing '}' at the end of a block statement.
S7.5.2_A1.5

PASS Expected parsing failure

TEST COMPLETE
